ExcelCreatorInterface added Fix mail sending(makes infinity loop and don't send mail. Why ? dick knows...)

+++ src/main/java/com/example/pizza_delivery/model/ZakazEntity.java
@@ -4,10 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.Cascade;
 
 import javax.persistence.*;
-import java.util.ArrayList;
 import java.util.List;
 
 @Entity
@@ -19,7 +17,7 @@ import java.util.List;
 public class ZakazEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @Column(name = "status")
     private String status;
